backend/requirements.txt
fastapi
aiosqlite
sqlalchemy
python-socketio